requirements.txt
googlesearch-python
langchain-community
langchain-core
langchain
sentence-transformers
faiss-cpu
